requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
mysql-connector-python

